Modificando casting de inteiros para floats nos coeficientes do polinômio do exercicio7

+++ Exercicios/exercicio7.py
@@ -2,7 +2,7 @@
 #equação do segundo grau a^x² + bx + c = 0
 
 def le_termo_dominante():
-    a = int(input("Digite o coeficiente do termo quadrático: "))
+    a = float(input("Digite o coeficiente do termo quadrático: "))
 
     while(a == 0):
         a = int(input("O coeficiente dominante não pode ser nulo. Entre outro coeficiente: "))
@@ -14,11 +14,11 @@ negativo = True
 
 while(negativo):
     a = le_termo_dominante()
-    b = int(input("Digite o coeficiente do termo 'x': "))
-    c = int(input("Digite o termo independente: "))
+    b = float(input("Digite o coeficiente do termo 'x': "))
+    c = float(input("Digite o termo independente: "))
 
     deltha = b**2 - 4*a*c
-    
+
     if deltha < 0:
         print("Discriminante negativo. Entre outros valores para a, b e c")
     else:
